Implement 'show' action
Accesses MIT Recreation's scheduling system when prompted to check availability of Z-Center courts. MIT only allows scheduling up to one day in advance, so the bot can check the schedule for today and tomorrow. Sets the image's timezone to US/Eastern to avoid confusion over the meaning of today and tomorrow.