Make login more robust, and avoid login when checking availability
Instead of clicking the login button and immediately proceeding, wait for elements to appear that are only seen after logging in. This ensures that the session cookie is set before the bot tries to make reservations. Also, it turns out a session cookie isn't necessary to check court availability, so the bot no longer passes the session cookie along with availability requests.