Fix deploy artifact copying
We want to copy the contents of the artifact directory over the contents of the courtbot directory, creating it if it doesn't exist. Before, this was creating an artifact directory inside the courtbot directory if the courtbot directory already existed.

Also remove reference to Docker Cloud. We only need Travis now!